pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
mypy = "^1.19.1"

[project]
name = "api"
version = "0.1.0"
description = "API for petergof science rag"
authors = [{ name = "andrey-blight", email = "kizhinov05@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aiobotocore (>=3.1.3,<4.0.0)",
    "boto3 (>=1.42.49,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "greenlet (>=3.3.1,<4.0.0)",
    "openai (>=2.17.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
